feat: refactor InsightsSection and BlogDetails components to use global state, add data structure for chapters and blog posts
