+++ src/store/store.ts
@@ -1,258 +1,8 @@
 import { create } from 'zustand'
-const Chapters = [
-  {
-    title: "Chapter 1: Foundations of UI/UX Design",
-    lessons: [
-      {
-        id: 1,
-        nameLesson: "Introduction to UI/UX",
-        Time_Of_lesson: 12,
-        link: "https://www.youtube.com/embed/zHAa-m16NGk"
-      },
-      {
-        id: 2,
-        nameLesson: "Understanding  Design",
-        Time_Of_lesson: 20,
-        link: "https://www.youtube.com/embed/Hbn1SMUs-hs"
-      },
-      {
-        id: 3,
-        nameLesson: "Design Thinking Process",
-        Time_Of_lesson: 25,
-        link: "https://www.youtube.com/embed/cJSDKVOW2cA"
-      },
-      {
-        id: 4,
-        nameLesson: "Principles of Good UX",
-        Time_Of_lesson: 18,
-        link: "https://www.youtube.com/embed/EUHPgbHcKy8"
-      },
-    ],
-    Number_Of_lessons: 4,
-    Time_Of_lessons: 75,
-  },
-  {
-    title: "Chapter 2: Foundations of UI/UX Design",
-    lessons: [
-      {
-        id: 1,
-        nameLesson: "Introduction to UI/UX",
-        Time_Of_lesson: 12,
-        link: "https://www.youtube.com/embed/zHAa-m16NGk"
-      },
-      {
-        id: 2,
-        nameLesson: "Understanding  Design",
-        Time_Of_lesson: 20,
-        link: "https://www.youtube.com/embed/Hbn1SMUs-hs"
-      },
-      {
-        id: 3,
-        nameLesson: "Design Thinking Process",
-        Time_Of_lesson: 25,
-        link: "https://www.youtube.com/embed/cJSDKVOW2cA"
-      },
-      {
-        id: 4,
-        nameLesson: "Principles of Good UX",
-        Time_Of_lesson: 18,
-        link: "https://www.youtube.com/embed/EUHPgbHcKy8"
-      },
-    ],
-    Number_Of_lessons: 4,
-    Time_Of_lessons: 75,
-  },
-  {
-    title: "Chapter 3: Foundations of UI/UX Design",
-    lessons: [
-      {
-        id: 1,
-        nameLesson: "Introduction to UI/UX",
-        Time_Of_lesson: 12,
-        link: "https://www.youtube.com/embed/zHAa-m16NGk"
-      },
-      {
-        id: 2,
-        nameLesson: "Understanding  Design",
-        Time_Of_lesson: 20,
-        link: "https://www.youtube.com/embed/Hbn1SMUs-hs"
-      },
-      {
-        id: 3,
-        nameLesson: "Design Thinking Process",
-        Time_Of_lesson: 25,
-        link: "https://www.youtube.com/embed/cJSDKVOW2cA"
-      },
-      {
-        id: 4,
-        nameLesson: "Principles of Good UX",
-        Time_Of_lesson: 18,
-        link: "https://www.youtube.com/embed/EUHPgbHcKy8"
-      },
-    ],
-    Number_Of_lessons: 4,
-    Time_Of_lessons: 75,
-  },
-  {
-    title: "Chapter 4: Foundations of UI/UX Design",
-    lessons: [
-      {
-        id: 1,
-        nameLesson: "Introduction to UI/UX",
-        Time_Of_lesson: 12,
-        link: "https://www.youtube.com/embed/zHAa-m16NGk"
-      },
-      {
-        id: 2,
-        nameLesson: "Understanding  Design",
-        Time_Of_lesson: 20,
-        link: "https://www.youtube.com/embed/Hbn1SMUs-hs"
-      },
-      {
-        id: 3,
-        nameLesson: "Design Thinking Process",
-        Time_Of_lesson: 25,
-        link: "https://www.youtube.com/embed/cJSDKVOW2cA"
-      },
-      {
-        id: 4,
-        nameLesson: "Principles of Good UX",
-        Time_Of_lesson: 18,
-        link: "https://www.youtube.com/embed/EUHPgbHcKy8"
-      },
-    ],
-    Number_Of_lessons: 4,
-    Time_Of_lessons: 75,
-  },
-];
-const AllMentors = [
-  {
-    id: 0,
-    img: "https://cdn.prod.website-files.com/681546135f34d96fc2aab532/68745af6ce0804310f4e5d48_pexels-zen-chung-5745170_enhanced%20(1)-p-1600.webp",
-    namePosition: "Web Developer",
-    name: "David Tan",
-    socialMedia: [
-      {
-        link: "www.google.com",
-        icon: "https://cdn.prod.website-files.com/67fdceadc4bf2fac05c7defb/6823527eea8a09e021553a63_twitter.svg",
-      },
-      {
-        link: "www.google.com",
-        icon: "https://cdn.prod.website-files.com/67fdceadc4bf2fac05c7defb/6823527e1a24265b12ffa3ed_linkedin.svg",
-      },
-      {
-        link: "www.google.com",
-        icon: "https://cdn.prod.website-files.com/67fdceadc4bf2fac05c7defb/6823527e723ce1c55a9a7417_instagram.svg",
-      },
-    ],
-  },
-  {
-    id: 1,
+import { AllMentors, blogPosts, Chapters } from './data'
 
-    img: "https://cdn.prod.website-files.com/681546135f34d96fc2aab532/68745af6ce0804310f4e5d48_pexels-zen-chung-5745170_enhanced%20(1)-p-1600.webp",
-    namePosition: "Web Developer",
-    name: "David Tan",
-    socialMedia: [
-      {
-        link: "www.google.com",
-        icon: "https://cdn.prod.website-files.com/67fdceadc4bf2fac05c7defb/6823527eea8a09e021553a63_twitter.svg",
-      },
-      {
-        link: "www.google.com",
-        icon: "https://cdn.prod.website-files.com/67fdceadc4bf2fac05c7defb/6823527e1a24265b12ffa3ed_linkedin.svg",
-      },
-      {
-        link: "www.google.com",
-        icon: "https://cdn.prod.website-files.com/67fdceadc4bf2fac05c7defb/6823527e723ce1c55a9a7417_instagram.svg",
-      },
-    ],
-  },
-  {
-    id: 2,
-
-    img: "https://cdn.prod.website-files.com/681546135f34d96fc2aab532/68745af6ce0804310f4e5d48_pexels-zen-chung-5745170_enhanced%20(1)-p-1600.webp",
-    namePosition: "Web Developer",
-    name: "David Tan",
-    socialMedia: [
-      {
-        link: "www.google.com",
-        icon: "https://cdn.prod.website-files.com/67fdceadc4bf2fac05c7defb/6823527eea8a09e021553a63_twitter.svg",
-      },
-      {
-        link: "www.google.com",
-        icon: "https://cdn.prod.website-files.com/67fdceadc4bf2fac05c7defb/6823527e1a24265b12ffa3ed_linkedin.svg",
-      },
-      {
-        link: "www.google.com",
-        icon: "https://cdn.prod.website-files.com/67fdceadc4bf2fac05c7defb/6823527e723ce1c55a9a7417_instagram.svg",
-      },
-    ],
-  },
-  {
-    id: 3,
-
-    img: "https://cdn.prod.website-files.com/681546135f34d96fc2aab532/68745af6ce0804310f4e5d48_pexels-zen-chung-5745170_enhanced%20(1)-p-1600.webp",
-    namePosition: "Web Developer",
-    name: "David Tan",
-    socialMedia: [
-      {
-        link: "www.google.com",
-        icon: "https://cdn.prod.website-files.com/67fdceadc4bf2fac05c7defb/6823527eea8a09e021553a63_twitter.svg",
-      },
-      {
-        link: "www.google.com",
-        icon: "https://cdn.prod.website-files.com/67fdceadc4bf2fac05c7defb/6823527e1a24265b12ffa3ed_linkedin.svg",
-      },
-      {
-        link: "www.google.com",
-        icon: "https://cdn.prod.website-files.com/67fdceadc4bf2fac05c7defb/6823527e723ce1c55a9a7417_instagram.svg",
-      },
-    ],
-  },
-  {
-    id: 4,
-
-    img: "https://cdn.prod.website-files.com/681546135f34d96fc2aab532/68745af6ce0804310f4e5d48_pexels-zen-chung-5745170_enhanced%20(1)-p-1600.webp",
-    namePosition: "Web Developer",
-    name: "David Tan",
-    socialMedia: [
-      {
-        link: "www.google.com",
-        icon: "https://cdn.prod.website-files.com/67fdceadc4bf2fac05c7defb/6823527eea8a09e021553a63_twitter.svg",
-      },
-      {
-        link: "www.google.com",
-        icon: "https://cdn.prod.website-files.com/67fdceadc4bf2fac05c7defb/6823527e1a24265b12ffa3ed_linkedin.svg",
-      },
-      {
-        link: "www.google.com",
-        icon: "https://cdn.prod.website-files.com/67fdceadc4bf2fac05c7defb/6823527e723ce1c55a9a7417_instagram.svg",
-      },
-    ],
-  },
-  {
-    id: 5,
-
-    img: "https://cdn.prod.website-files.com/681546135f34d96fc2aab532/68745af6ce0804310f4e5d48_pexels-zen-chung-5745170_enhanced%20(1)-p-1600.webp",
-    namePosition: "Web Developer",
-    name: "David Tan",
-    socialMedia: [
-      {
-        link: "www.google.com",
-        icon: "https://cdn.prod.website-files.com/67fdceadc4bf2fac05c7defb/6823527eea8a09e021553a63_twitter.svg",
-      },
-      {
-        link: "www.google.com",
-        icon: "https://cdn.prod.website-files.com/67fdceadc4bf2fac05c7defb/6823527e1a24265b12ffa3ed_linkedin.svg",
-      },
-      {
-        link: "www.google.com",
-        icon: "https://cdn.prod.website-files.com/67fdceadc4bf2fac05c7defb/6823527e723ce1c55a9a7417_instagram.svg",
-      },
-    ],
-  },
-];
 export const useStore = create(() => ({
-  Chapters: Chapters,
-  AllMentors: AllMentors,
+  Chapters,
+  AllMentors,
+  blogPosts
 }))
